+++ src/components/myModal/index.js
@@ -11,6 +11,40 @@ const defaultProps = {
 	left: 1
 }
 
+/*
+*
+example
+*
+*/
+/*
+<MyModal
+	visible = { this.state.visible }
+	top = {0.5}
+	left = {0.5}
+	contentStyle = {{ flex: 3 }}
+	onRequestClose = { this._setModalVisible.bind(this, false) }>
+	<View style = {{ flex: 1, width: width - 20, height: height - 100, padding: 5, borderRadius: 5, backgroundColor: 'white' }}>
+		<View style = { styles.stickyBottom }>
+			<View style = { styles.row }>
+				<Button
+					onPress = { () => this.setState({ idProduct: null,
+						barcode: null,
+						product: null,
+						cost: null,
+						price: null,
+						quantity: null })}
+					name = 'Clear'/>
+
+				<Button
+					onPress = { this._addProduct.bind(this) }
+					name = 'Add'/>
+			</View>
+		</View>
+	</View>
+</MyModal>
+*/
+/**/
+
 class MyModal extends React.Component {
 	render() {
 		return (