+++ src/components/touchable/index.js
@@ -11,6 +11,19 @@ const defaultProps = {
 	style: {}
 }
 
+/*
+*
+example
+*
+*/
+/*
+<Touchable
+	onPress = { this._deleteCategory.bind(this, content) }>
+	<Text> X </Text>
+</Touchable>
+*/
+/**/
+
 class Touchable extends React.Component {
 	render() {
 		if(Platform.OS == 'android' && Platform.Version >= 21) {
@@ -19,7 +32,7 @@ class Touchable extends React.Component {
 					ref = { this.props.ref }
 					background = { TouchableNativeFeedback.Ripple(this.props.pressColor, true) }
 					onPress = { this.props.onPress == null ? null : this.props.onPress.bind(this) } >
-					<View style = { this.props.style }>
+					<View style = {[ this.props.style, { flex: 1 }]}>
 						{ this.props.children }
 					</View>
 				</TouchableNativeFeedback>
@@ -29,7 +42,7 @@ class Touchable extends React.Component {
 		return (
 			<TouchableOpacity
 				ref = { this.props.ref }
-				style = { this.props.style }
+				style = {[ this.props.style, { flex: 1 }]}
 				onPress = { this.props.onPress == null ? null : this.props.onPress.bind(this) } >
 				{ this.props.children }
 			</TouchableOpacity>